moe: explicit dtype cast on weighted before index_add_
torch.Tensor.index_add_ requires source and self to share dtype.
weights_e is fp32 from topk_weights and unsqueeze(1) makes it 2-D, which
triggers PyTorch's promotion rules during the multiply with the bf16
down_out cast (a 2-D tensor on the right will promote, unlike the
original code's 0-D scalar weight). Explicit .to(x.dtype) on weights_e
keeps the multiply in bf16, matching the original code's
accumulation precision.

+++ src/ds4_hybrid_quant/vllm_patches/moe_method.py
@@ -384,7 +384,10 @@ class Iq2XxsQ2KFusedMoEMethod(FusedMoEMethodBase):
                 mid_qs, mid_d, mid_bsums,
             )                                                   # (M_e, hidden)
 
-            weighted = down_out.to(x.dtype) * weights_e.unsqueeze(1)
+            # Cast both factors to x.dtype so index_add_'s dtype check
+            # passes — and to match the original code's bf16-throughout
+            # accumulation precision.
+            weighted = down_out.to(x.dtype) * weights_e.to(x.dtype).unsqueeze(1)
             out.index_add_(0, tokens_e, weighted)
 
         # DS4_HDUMP: hidden-state dump for layer-by-layer comparison vs ds4
